Todo creating ui package and moving relevant classes to it
